% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoCalcs_bootystrapper_uc.R
\name{doCalcs_uc}
\alias{doCalcs_uc}
\title{Take bootstrap output and return summary data for all annual report/web table for Upper and Mid Columbia River fish}
\usage{
doCalcs_uc(crt, target, css_group, makefile = "y", ...)
}
\arguments{
\item{crt}{Bootstrap output as input file here.}

\item{target}{The name of the original input file (eg. SR HCH 2015 MCCA).}

\item{css_group}{CSS group from the CSSGroups_LookupTable.}

\item{makefile}{Save bootstrap output in CSSOUTPUT in SQL server, append parameter output in CSSREPORT in SQL server, and make parameter output in csv file in working directory. Default is 'y'.}
}
\value{
Survivals,detection, adult counts, and SARs...
}
\description{
Take bootstrap output and return summary data for all annual report/web table for Upper and Mid Columbia River fish
}
\examples{
ans<- doCalcs_uc(crt, target='CR_USK_2015_OKAN_MCJ', css_group='OKSR', makefile='y')
round(ans, 3)
print(format(ans, scientific = F))
}
