% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/booty_utility.R
\name{format_dat}
\alias{format_dat}
\title{Import and format data for ready to use by \code{surv_calc()}}
\usage{
format_dat(file_name, mig_yr = "auto", wgt = "n")
}
\arguments{
\item{file_name}{File path where the input csv file is stored.}

\item{mig_yr}{Juvenile migration year.}

\item{wgt}{"y" if using weighted sampling probability. User will be prompted to enter the amount of intergrated and segregated fish.}
}
\value{
Capture history and indicators for adult return.
}
\description{
Import and format data for ready to use by \code{surv_calc()}
}
\examples{
detect_data<- format_dat('C:/Users/bobbyhsu/Documents/Temp/SR HCH 2014 MCCA.csv', mig_yr= 2014, wgt= 'n')

}
