% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/booty_utility.R
\name{mark_calc}
\alias{mark_calc}
\title{Estimate survivals/detection using likelihood method, with logit link, and tally adult returns}
\usage{
mark_calc(ch, i, nocc, wt, phi_p_only, ...)
}
\arguments{
\item{ch}{Input file made by \code{format_dat()} function.}

\item{i}{Iteration number used by the bootstrap function \code{bootystrapper()}.}

\item{nocc}{Total detection events including the trawl.}

\item{wt}{Indicates whether to weight the sampling probability (bootstrap only).}

\item{phi_p_only}{Option to only calculate survivals and detection and not do the adult counts.}
}
\value{
Survivals, detection and returing adult counts
}
\description{
Estimate survivals/detection using likelihood method, with logit link, and tally adult returns
}
\examples{
mark_calc(detect_data, i= 1, nocc= 6, wt= 'n', phi_p_only= 'y')

}
