% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/booty_utility.R
\name{mark_dat}
\alias{mark_dat}
\title{Make an ".inp" file for packages RMark.}
\usage{
mark_dat(ch)
}
\arguments{
\item{ch}{Input file containing capture history (from \code{format_dat()} function).}
}
\value{
RMark data file
}
\description{
Make an ".inp" file for packages RMark.
}
\examples{
m_data<- mark_dat(ch)

}
