% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/booty_utility.R
\name{marray}
\alias{marray}
\title{Construct m-ij array}
\usage{
marray(CH, n_occ)
}
\arguments{
\item{CH}{Input file made by \code{format_dat()} function.}

\item{n_occ}{Total detection events including the trawl.}
}
\value{
m-ij array
}
\description{
Construct m-ij array
}
\examples{
marray(detect_data, n_occ=8)

}
