% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/booty_utility.R
\name{marray_wtd}
\alias{marray_wtd}
\title{Construct m-ij array with weighted probability}
\usage{
marray_wtd(CH, n_occ)
}
\arguments{
\item{CH}{Input file made by \code{format_dat()} function.}

\item{n_occ}{Total detection events including the trawl.}
}
\value{
m-ij array
}
\description{
Construct m-ij array with weighted probability
}
\details{
It needs pre-assigned weights to calculate weighted m-ij. If all weights are the same, the results are the same as the unweighted m-ij.
}
\examples{
# pre-assign weights for intergrated and segregated populations
intgr<- 234012/(234012+813874) # 2014 CW
segr <- 813874/(234012+813874) # 2014 AD
# make input file, specify weighted sampling
wgt_data<- format_dat('C:/Users/bobbyhsu/Documents/Temp/SR HCH 2014 MCCA.csv', wgt='y')
marray(wgt_data, n_occ= 8)

}
