% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fake_data_cjs.R
\name{siml_cjs}
\alias{siml_cjs}
\title{data simulation for single release, diff surv for int and seg groups}
\usage{
siml_cjs(big_phi, big_p, mrkd, remv, n_occ, intgr, surv_diff, grp_t,
  adu_rtn)
}
\arguments{
\item{big_phi}{Set reach survivals.}

\item{big_p}{Set detection probabilities for each reach.}

\item{mrkd}{Indicates amount of fish released.}

\item{remv}{Indicates the portion of fish removed (transported or died).}

\item{n_occ}{Number of events/detection.}

\item{intgr}{Portion of "integrated" group. Integrated + Segregated = 1.}

\item{surv_diff}{Difference in survival between the intergrated and segregated groups.}

\item{grp_t}{Portion of transported group. Group T + Group R = 1}

\item{adu_rtn}{Adult return rate.}
}
\value{
Simulated detection history and adult counts. Bootstrap ready.
}
\description{
data simulation for single release, diff surv for int and seg groups
}
\examples{
n_occ<- 8
mrkd<- 5000
phi_real<- rep(0.8, n_occ-1) # set survival
p_real<- rep(0.45, n_occ-1) # set detection
remv<- 0.01 # portion removed at occ 2 and 3
intgr<- 0.23
surv_diff<- 0.07
grp_t<- 0.7
adu_rtn<- 0.02
big_phi<- matrix(phi_real, ncol= n_occ-1, nrow= mrkd)
big_p<- matrix(p_real, ncol= n_occ-1, nrow= mrkd)

ch<- siml_cjs(big_phi, big_p, mrkd, remv, n_occ, intgr, surv_diff, grp_t, adu_rtn)

}
