% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/booty_utility.R
\name{surv_calc}
\alias{surv_calc}
\title{Estimate survivals, detections, and tally adult returns}
\usage{
surv_calc(ch, i, nocc, wt, wt_i)
}
\arguments{
\item{ch}{Input file made by \code{format_dat()} function.}

\item{i}{Iteration number used by the bootstrap function \code{bootystrapper()}.}

\item{wt}{Indicates whether to weight the sampling probability.}

\item{wt_i}{Indicates whether to calculate the original estimates using weighted probability.}
}
\value{
Survivals, detection and returing adult counts
}
\description{
Estimate survivals, detections, and tally adult returns
}
\examples{
for(i in 1:10){
  results<- surv_calc(detect_data, i, nocc=8, wt='y', wt_i='y')
}
results

}
