% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/booty_utility.R
\name{bootystrapper}
\alias{bootystrapper}
\title{Bootstrap using surv_calc and organize output}
\usage{
bootystrapper(d, fn, iter, n_occ, wgt, wgt_init)
}
\arguments{
\item{d}{Input file made by \code{format_dat()}.}

\item{fn}{Function to run the bootstrap on.}

\item{iter}{Amount of bootstrap iterations.}

\item{n_occ}{Total detection events including the trawl.}

\item{wgt}{Indicates whether to weight the sampling probability.}

\item{wgt_int}{Indicates whether to calculate the original estimates using weighted probability.}
}
\value{
Estimates in a data frame with original estimate as the first row and bootstrap results in the remaining rows.
}
\description{
Bootstrap using surv_calc and organize output
}
\examples{
out<- bootystrapper(detect_data, surv_calc, iter=100, n_occ=8, wgt='n', wgt_init='n')
head(out)

}
