% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/booty_utility.R
\name{bootystrapper}
\alias{bootystrapper}
\title{Bootstrap using surv_calc and organize output}
\usage{
bootystrapper(d, fn, iter, wgt, wgt_init, phi_p_only = "n", fpc = "y",
  match_bt4 = "y", logit_link = "n", ...)
}
\arguments{
\item{d}{Input file made by \code{format_dat()}.}

\item{fn}{Function to run the bootstrap on.}

\item{iter}{Amount of bootstrap iterations.}

\item{wgt}{Indicates whether to weight the sampling probability.}

\item{wgt_init}{Indicates whether to calculate the original estimates using weighted probability.}

\item{phi_p_only}{Indicate to turn off the phi_p_only option in \code{curv_calc()}. Default is no ("n").}

\item{fpc}{Indicate to turn off the finite population correction option in \code{curv_calc()}. Default is yes ("y").}

\item{match_bt4}{Indicate to turn off the match_bt4 option in \code{curv_calc()}. The default here is yes ("y").}

\item{logit_link}{Indicate to use "RMark" or "marked" and estimate using logit link. The default here is none ("n").}
}
\value{
Estimates in a data frame with original estimate as the first row and bootstrap results in the remaining rows.
}
\description{
Bootstrap using surv_calc and organize output
}
\examples{
out<- bootystrapper(detect_data, surv_calc, iter= 100, n_occ= 8, wgt= 'n', wgt_init= 'n')
head(out)

}
