% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/booty_utility.R
\name{format_dat}
\alias{format_dat}
\title{Import and format data for ready surv_calc}
\usage{
format_dat(file_name, wgt)
}
\arguments{
\item{file_name}{File path where the input csv file is stored.}

\item{wgt}{'y' if using weighted sampling probability.}
}
\value{
Capture history and indicators for adult return.
}
\description{
Import and format data for ready surv_calc
}
\examples{
detect_data<- format_dat('C:/Users/bobbyhsu/Documents/Temp/SR HCH 2014 MCCA.csv', wgt='n')

}
